+++ react-capstone/src/Components/Home/Home.jsx
@@ -4,9 +4,7 @@ import logo from '../Assets/logo.png'
 export default class Home extends React.Component{
     render(){
         return(
-            <div>
-                <img src={logo} alt="logo" align="center"/>
-            </div>
+            <img src={logo} alt="logo" align="center"/>
         )
     }
 }
